uint32 -> uint16

+++ malle/bimap/bimap_test.go
@@ -11,27 +11,27 @@ func TestBimap(t *testing.T) {
 		t.Errorf("New() should return a map with Size 0, got %d", m.Size())
 	}
 
-	err := m.Add("", uint32(1))
+	err := m.Add("", uint16(1))
 	if err != ErrEmptyString {
 		t.Errorf("Bimap.Add() empty string should fail with ErrEmptyString")
 	}
 
-	err = m.Add("a", uint32(1))
+	err = m.Add("a", uint16(1))
 	if err != nil {
 		t.Errorf("Bimap.Add() failed with %v", err)
 	}
 
 	i, ok := m.FindByStr("a")
-	if !ok || i != uint32(1) {
+	if !ok || i != uint16(1) {
 		t.Errorf("FindByStr(\"a\") => %d, %v, want 1, true", i, ok)
 	}
 
-	s, ok := m.FindByInt(uint32(1))
+	s, ok := m.FindByInt(uint16(1))
 	if !ok || s != "a" {
 		t.Errorf("FindByStr(1) => %s, %v, want \"a\", true", s, ok)
 	}
 
-	err = m.Add("b", uint32(99))
+	err = m.Add("b", uint16(99))
 	if err != nil {
 		t.Errorf("Bimap.Add() failed with %v", err)
 	}
@@ -55,17 +55,17 @@ func TestBimap(t *testing.T) {
 		t.Errorf("Bimap.RemoveByStr(\"a\") didn't remove map entry")
 	}
 
-	_, ok = m.FindByInt(uint32(1))
+	_, ok = m.FindByInt(uint16(1))
 	if ok {
 		t.Errorf("Bimap.RemoveByStr(\"a\") didn't remove map entry")
 	}
 
-	err = m.RemoveByInt(uint32(99))
+	err = m.RemoveByInt(uint16(99))
 	if err != nil {
 		t.Errorf("Bimap.RemoveByStr(99) => %v; want no error", err)
 	}
 
-	_, ok = m.FindByInt(uint32(99))
+	_, ok = m.FindByInt(uint16(99))
 	if ok {
 		t.Errorf("Bimap.RemoveByInt(99) didn't remove map entry")
 	}
@@ -83,7 +83,7 @@ func TestBimap(t *testing.T) {
 func TestGrowingBimap(t *testing.T) {
 	m := New(2)
 	for i := 0; i < 1000; i++ {
-		err := m.Add(fmt.Sprintf("n:%d", i), uint32(i))
+		err := m.Add(fmt.Sprintf("n:%d", i), uint16(i))
 		if err != nil {
 			t.Fatal(err)
 		}
